requirements.txt
Flask
psycopg2-binary
SQLAlchemy
alembic
python-dotenv